Builder add validations

+++ src/main/java/org/bluepigeon/admin/dao/BuyerDAO.java
@@ -880,29 +880,42 @@ public class BuyerDAO {
 	 * @return response message
 	 */
 	public ResponseMessage deleteSecondaryBuyerById(int id){
-		String hql = "from BuyerDocuments where buyer.id = :id";
-		HibernateUtil hibernateUtil = new HibernateUtil();
-		Session session = hibernateUtil.openSession();
-		Query query = session.createQuery(hql);
-		query.setParameter("id", id);
-		List<BuyerDocuments> result = query.list();
-		session.beginTransaction();
-		for(BuyerDocuments buyerDocuments :result) {
-			session.delete(buyerDocuments);
-		}
-		session.getTransaction().commit();
-		session.close();
-		
-		Buyer buyer = new Buyer();
-		buyer.setId(id);
 		ResponseMessage resp = new ResponseMessage();
-		Session session2 = hibernateUtil.openSession();
-		session2.beginTransaction();
-		session2.delete(buyer);
-		session2.getTransaction().commit();
-		session2.close();
-		resp.setMessage("Buyer deleted successfully.");
-		resp.setStatus(1);
+		String hql = "from BuyerDocuments where buyer.id = :id";
+		String bhql = "from Buyer where id = :id";
+		HibernateUtil hibernateUtil = new HibernateUtil();
+		
+		Session bsession = hibernateUtil.openSession();
+		Query bquery = bsession.createQuery(bhql);
+		bquery.setParameter("id", id);
+		List<Buyer> bresult = bquery.list();
+		if(bresult.get(0).getIsPrimary()) {
+			resp.setStatus(0);
+			resp.setMessage("You can not delete primary buyer");
+		} else {
+		
+			Session session = hibernateUtil.openSession();
+			Query query = session.createQuery(hql);
+			query.setParameter("id", id);
+			List<BuyerDocuments> result = query.list();
+			session.beginTransaction();
+			for(BuyerDocuments buyerDocuments :result) {
+				session.delete(buyerDocuments);
+			}
+			session.getTransaction().commit();
+			session.close();
+			
+			Buyer buyer = new Buyer();
+			buyer.setId(id);
+			Session session2 = hibernateUtil.openSession();
+			session2.beginTransaction();
+			session2.delete(buyer);
+			session2.getTransaction().commit();
+			session2.close();
+			resp.setMessage("Buyer deleted successfully.");
+			resp.setStatus(1);
+		}
+		
 		return resp;
 	}
 		
@@ -986,6 +999,24 @@ public class BuyerDAO {
 			response.setMessage("Flat status updated");
 		}
 		return response;
+	}
+	//To display total buyers on dash board
+	/**
+	 * Get total count of buyers by builder id
+	 * @author pankaj
+	 * @param builderId
+	 * @return count of buyers
+	 */
+	public Long getTotalBuyers(int builderId){
+		Long totalBuyers = (long) 0;
+		String hql = "select COUNT(*) from Buyer where builder.id = :builder_id and is_deleted = 0";
+		HibernateUtil hibernateUtil = new HibernateUtil();
+		Session session = hibernateUtil.openSession();
+		Query query = session.createQuery(hql);
+		query.setInteger("builder_id", builderId);
+		totalBuyers = (Long) query.uniqueResult();
+		
+		return totalBuyers;
 	}
 	
 }
